refactor(#394): extract inline `mod tests` into sibling _tests.rs (29 files)
The whole-repo Static Checks validator (scripts/validate.sh rule 5, issue #394)
bans inline `#[cfg(test)] mod tests { ... }` in production files — tests live in
<module>_tests.rs wired via `#[cfg(test)] #[path = "..."] mod tests;`. 29 files
across 9 crates still violated it, keeping the Test workflow red (on develop too).

Behavior-preserving move: each `mod tests` body extracted verbatim (dedented one
level) into its sibling _tests.rs and re-included via #[path]. No test changed.
di_output_options keeps its separate `duplicate_label_tests` module (not the
banned `mod tests` name). chain_row_wiring_tests.rs was a stale unreferenced
orphan (dead symbols) — overwritten with the current wired test body.

cargo check --workspace --tests: clean, 0 warnings.
