+++ crates/project/src/channel_mode_conv.rs
@@ -67,62 +67,5 @@ impl TryFrom<ChannelMode> for ChainOutputMode {
 // `TryFrom<ChannelMode> for ChainOutputMode`), kept above.
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use domain::io_binding::ChannelMode;
-
-    // ── From<ChannelMode> for ChainInputMode ────────────────────────────────
-
-    #[test]
-    fn channel_mode_mono_converts_to_chain_input_mono() {
-        assert_eq!(
-            ChainInputMode::from(ChannelMode::Mono),
-            ChainInputMode::Mono
-        );
-    }
-
-    #[test]
-    fn channel_mode_stereo_converts_to_chain_input_stereo() {
-        assert_eq!(
-            ChainInputMode::from(ChannelMode::Stereo),
-            ChainInputMode::Stereo
-        );
-    }
-
-    #[test]
-    fn channel_mode_dual_mono_converts_to_chain_input_dual_mono() {
-        assert_eq!(
-            ChainInputMode::from(ChannelMode::DualMono),
-            ChainInputMode::DualMono
-        );
-    }
-
-    // ── TryFrom<ChannelMode> for ChainOutputMode ────────────────────────────
-
-    #[test]
-    fn channel_mode_mono_converts_to_chain_output_mono() {
-        assert_eq!(
-            ChainOutputMode::try_from(ChannelMode::Mono).unwrap(),
-            ChainOutputMode::Mono
-        );
-    }
-
-    #[test]
-    fn channel_mode_stereo_converts_to_chain_output_stereo() {
-        assert_eq!(
-            ChainOutputMode::try_from(ChannelMode::Stereo).unwrap(),
-            ChainOutputMode::Stereo
-        );
-    }
-
-    #[test]
-    fn channel_mode_dual_mono_fails_for_chain_output_with_clear_error() {
-        let err = ChainOutputMode::try_from(ChannelMode::DualMono).unwrap_err();
-        assert_eq!(err.source, ChannelMode::DualMono);
-        assert!(
-            err.to_string().contains("dual-mono"),
-            "error message should mention dual-mono: {}",
-            err
-        );
-    }
-}
+#[path = "channel_mode_conv_tests.rs"]
+mod tests;
